+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 ### 🚀 Features
 
 - `CustomTransferMap` elements created by combining multiple other elements will now reflect that in their `name` attribute (see #100) (@jank324)
-- Add a new class method for `ParticleBeam` to generate 3D uniformly distributed ellipsoidal beam. (see #146) (@cr-xu)
+- Add a new class method for `ParticleBeam` to generate a 3D uniformly distributed ellipsoidal beam (see #146) (@cr-xu, @jank324)
 
 ### 🐛 Bug fixes